requirements.txt
streamlit
python-docx
regex
google-generativeai
chromadb
sentence-transformers
python-dotenv
pdfplumber
